handle seg in big map

+++ raycasting/doors.c
@@ -6,7 +6,7 @@
 /*   By: sodahani <sodahani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 22:48:03 by sodahani          #+#    #+#             */
-/*   Updated: 2025/08/20 15:39:50 by sodahani         ###   ########.fr       */
+/*   Updated: 2025/08/20 17:23:45 by sodahani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,17 +63,17 @@ void	update_doors(t_game *game)
 
 void	try_open_doors(t_game *game)
 {
-	int	px;
-	int	py;
+	float	px;
+	float	py;
 	int	i;
 
 
-    px = (int)(game->player_x / TILE_SIZE);
-    py = (int)(game->player_y / TILE_SIZE);
+    px = (game->player_x / TILE_SIZE);
+    py = (game->player_y / TILE_SIZE);
 	i = 0;
 	while (i < game->num_doors)
 	{
-		if (abs(game->doors[i].x - py) + abs(game->doors[i].y - px) == 1)
+		if (abs(game->doors[i].x - (int) py) + abs(game->doors[i].y - (int) px) == 1)
 		{
 			game->doors[i].is_open = !game->doors[i].is_open;
 			update_doors(game);
